fix: carry UI message payloads at pointer width, and bound what that reaches
The client's UI message bus declared its two payload slots as int:

    Base::SendMessage(DWORD, int left, int right, void*)
      -> struct MESSAGE { int left; int right; }
      -> UI_ResultReceiver -> UIMessageManager::Execute
      -> m_UIMessageFunction[msg](left, right, void_ptr)

Eleven call sites put a pointer in those slots - an MItem*, or the c_str() of
a std::string - and the handlers cast them straight back, with
`MItem* pItem = (MItem*)left;` and several `(char*)left`. int is 32 bits on
64-bit Windows, so every one of them truncated a heap pointer and sign
extended it back. Item renaming, the pet and custom nickname flows, the SMS
dialogs, the guild team registration and the christmas tree have all been
crash on use since the SDL port: `a6e21d2` rewrote the original `(int)ptr`
into `(int)(intptr_t)ptr`, which silences the x64 truncation error while
preserving the truncation.

This is the same defect the review recorded as C21 for DescriptorManager, in
a second and larger mechanism, and it is not among the 197 findings.

The payload path is now intptr_t end to end: the MESSAGE struct, SendMessage
and _SendMessage, the three function pointer typedefs, UI_ResultReceiver,
UIMessageManager::Execute, UI_MESSAGE_FUNCTION and every handler signature.
Unlike the KeyboardControl `long extra` in C23 - which is virtual, where a
missed override silently becomes an overload that never fires - this bus
dispatches through a function pointer table, so a handler left at the old
width is a hard compile error. That is what makes the change verifiable
rather than hopeful, and it earned its keep: the first pass covered four
textual signature variants and the build stopped on two handlers written with
a fifth.

TempInformation::Value1..4 are widened for the same reason. They are the
parking slot between sending a request and the server answering it, and
GCAddressListVerifyHandler and GCNicknameVerifyHandler cast them back to
char*, so leaving them int would have truncated the pointers again one step
later.

Making these paths reachable is the dangerous part of the change, because
three memory safety bugs sit behind them that the truncation was masking. An
adversarial review found all three and they are fixed here rather than left
for the crash to be replaced by something quieter:

  - C_VS_UI_SMS_MESSAGE::AddSendList copied a server supplied phone number
    into char[16] with wsprintf, which bounds output at 1024 bytes rather
    than at the destination. GCSMSAddressList reads that field with a BYTE
    length prefix, so a hostile or buggy server had roughly 240 bytes of
    stack past the buffer. It is a bounded copy now, and the `str+3` split
    and the `str[2]` read are guarded by a length check instead of assuming
    three characters.
  - Execute_UI_CHANGE_CUSTOM_NAMING strcpy'd the nickname into char[22].
    The editor's SetByteLimit(22) counts characters, not bytes, and the DBCS
    conversion emits up to two bytes each, so 22 Korean characters are 44.
    Bounded, into a buffer with room.
  - The SMS and nickname flows parked c_str() pointers in TempInformation
    across the round trip, and nothing keeps the owning dialog alive that
    long: closing it, or the ESC path through ClosePopupWindow, deletes it.
    TempInformation owns std::string copies now.

Also corrected: a slot comment on Excute_UI_USE_XMAS_TREE that described
left/right as "to" and "from" when left is the item pointer, and the stale
old-width declarations of UI_ResultReceiver in VS_UI/WinMain.cpp, which is
excluded from the build and so would not have failed until someone re-enabled
it.

One cost worth stating plainly: widening adds 158 new C4244 warnings where
handlers pass the now 64-bit payload into functions taking int, uint or
DWORD. Those handlers carry real integers and the narrowing is correct, but
the compiler cannot tell them apart from the ones that carried pointers. The
baseline is roughly 27,000 warnings, so this is noise level, and each new one
marks a genuine narrowing point in a bus whose bug class is exactly that.
Correcting my own earlier claim: I said no handler passes left/right to a
printf-family function. That was wrong - there are eight, including a
DEBUG_ADD_FORMAT with %d. They are harmless, because the MS x64 ABI gives
every integer vararg an eight byte slot and %d reads the low half without
losing sync, but the claim should not have been made without checking.

Verified: build/vs2022 and build/vs2022-asan Debug both build with 0 errors,
and 92 tests, 380 checks, 0 failed in both test trees. All of this is
executable-only code with no test path, so it is a build-verified regression
guard; the truncation was not reproduced against a live server.

+++ VS_UI/src/VS_UI_Base.cpp
@@ -134,7 +134,7 @@ void Base::SelectFont(font_id_t id)
 //
 // 
 //-----------------------------------------------------------------------------
-void Base::Init(CSpriteSurface *surface, void (*fp)(DWORD, int, int, void *))
+void Base::Init(CSpriteSurface *surface, void (*fp)(DWORD, intptr_t, intptr_t, void *))
 {
 	InitSurface(surface);
 	InitFont();
@@ -358,7 +358,7 @@ void Base::DispatchMessage()
 //
 // 
 //-----------------------------------------------------------------------------
-void Base::SendMessage(DWORD message, int left, int right, void * void_ptr)
+void Base::SendMessage(DWORD message, intptr_t left, intptr_t right, void * void_ptr)
 {
 	m_C_ui_result_receiver._SendMessage(message, left, right, void_ptr);
 }
