+++ VS_UI/src/header/VS_UI_ui_result_receiver.h
@@ -6,11 +6,19 @@
 #include "VS_UI_UIMessage.h"
 
 
+//
+// left and right are intptr_t, not int, because several senders put a pointer
+// in them - an MItem* or a c_str() - and int is 32 bits on 64-bit Windows, so
+// the pointer arrived at the handler with its top half gone. The dispatch runs
+// through a function pointer table, so every handler signature has to match
+// this width or the assignment is a compile error, which is what keeps the two
+// ends honest.
+//
 struct MESSAGE
 {
 	DWORD				message;
-	int				left;
-	int				right;
+	intptr_t			left;
+	intptr_t			right;
 	void *			void_ptr;
 };
 
@@ -38,19 +46,19 @@ class C_VS_UI_UI_RESULT_RECEIVER
 private:
 	C_MESSAGE_QUEUE			m_message_queue;
 
-	void (*m_fp_result_receiver)(DWORD, int, int, void *);
+	void (*m_fp_result_receiver)(DWORD, intptr_t, intptr_t, void *);
 
 public:
 	C_VS_UI_UI_RESULT_RECEIVER();
 	~C_VS_UI_UI_RESULT_RECEIVER();
 
-	void _SendMessage(DWORD message, int left = 0, int right = 0, void *void_ptr = NULL);
+	void _SendMessage(DWORD message, intptr_t left = 0, intptr_t right = 0, void *void_ptr = NULL);
 	void	_DispatchMessage();
 
 /*-----------------------------------------------------------------------------
   Set.
 -----------------------------------------------------------------------------*/
-	void SetResultReceiver(void (*fp)(DWORD, int, int, void *));
+	void SetResultReceiver(void (*fp)(DWORD, intptr_t, intptr_t, void *));
 
 #ifndef _LIB
 	int	GetMessageSize() const { return m_message_queue.Size(); }
